Modpack splitted on 3 modpack

+++ pack.ps1
@@ -1,7 +1,14 @@
-$outputFile = "$PSScriptRoot/LethalReina_Modpack.zip"
+param($mod)
+
+if (!(Test-Path "$PSScriptRoot/$mod")) {
+    Write-Error "Mod with name $mod not exists"
+    exit 1
+}
+
+$outputFile = "$PSScriptRoot/LethalReina_Modpack_$mod.zip"
 
 if (Test-Path $outputFile) {
     Remove-Item $outputFile
 }
 
-Compress-Archive -Path ("$PSScriptRoot/package/*", "$PSScriptRoot/README.md", "$PSScriptRoot/CHANGELOG.md") -DestinationPath $outputFile
+Compress-Archive -Path ("$PSScriptRoot/$mod/*", "$PSScriptRoot/README.md"") -DestinationPath $outputFile
